app/requirements.txt
Django~=4.2.1
pip==23.1.2
asgiref~=3.6.0
sqlparse~=0.4.4
setuptools==67.8.0
mysqlclient~=2.1.1
uwsgi~=2.0.21
djangorestframework~=3.14.0
cryptography~=41.0.1
Jinja2~=3.1.2
pytz~=2023.3
requests~=2.31.0
MarkupSafe~=2.1.3
certifi~=2023.5.7
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
uritemplate~=4.1.1
packaging~=23.1
inflection~=0.5.1
idna~=3.4
urllib3~=2.0.3
pycparser~=2.21
drf_yasg~=1.21.5
djangorestframework-simplejwt~=5.2.2